% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycount_package.R
\docType{package}
\name{mycount-package}
\alias{mycount}
\alias{mycount-package}
\title{mycount: Tools for grouped counts and incidence rates}
\description{
This package provides \code{count_bio()}, an extension of \code{dplyr::count()}
that can also compute person-time, incidence rates, and exact Poisson CIs.
}
\author{
\strong{Maintainer}: Vicky S. \email{weiyings@umich.edu}

}
\keyword{internal}
